finish random mode
