+++ src/staff/tripletNotesStaff.js
@@ -1,27 +1,9 @@
 import { Vex } from 'vexflow';
+import { annotate, drawStaff, getLocalStorage } from './staffHelpers';
 
 function renderTripletNotesStaff() {
-  const parentDiv = document.getElementById('notes-graph');
-
-  while (parentDiv.firstChild) {
-    parentDiv.removeChild(parentDiv.firstChild);
-  }
-
-  const { Renderer, Stave, StaveNote, Beam, Tuplet, Formatter, Annotation } =
-    Vex.Flow;
-  // Create an SVG renderer and attach it to the DIV element
-  const renderer = new Renderer(parentDiv, Renderer.Backends.SVG);
-
-  // Configure the rendering context.
-  renderer.resize(650, 200);
-  const context = renderer.getContext();
-  context.setFont('Arial', 10);
-
-  //Add stave
-  const stave = new Stave(25, 40, 600);
-
-  // Add a clef and time signature.
-  stave.addClef('percussion').addTimeSignature('4/4');
+  const { StaveNote, Beam, Tuplet, Formatter } = Vex.Flow;
+  const [context, stave] = drawStaff();
 
   let notes1;
   let notes2;
@@ -187,31 +169,6 @@ function renderTripletNotesStaff() {
   });
   //draw tuplets
   tuplets.forEach((t) => t.setContext(context).draw());
-  // Render the stave
-  stave.setContext(context).draw();
-
-  //helper function
-  function annotate(hand) {
-    return new Annotation(hand)
-      .setVerticalJustification(Annotation.VerticalJustify.BOTTOM)
-      .setFont('Arial', 14, 'bold');
-  }
-}
-
-function getLocalStorage() {
-  const regex = /^beat-/;
-  const beatArray = [];
-  // Loop through all items in localStorage and retrieve items that match the regular expression
-  for (let i = 0; i < localStorage.length; i++) {
-    const key = localStorage.key(i);
-    if (regex.test(key)) {
-      const value = JSON.parse(localStorage.getItem(key));
-      if (value.noteValue === 'triplet') {
-        beatArray.push(value);
-      }
-    }
-  }
-  return beatArray;
 }
 
 export default renderTripletNotesStaff;
